proof of concept
created LGChangeHandler, which will act as a delegate to multiple
NSTextFields.  It will receive textDidEndEditing notifications from
NSTextFields and update defaults.

I’m marking sections of saveAndClose: as Implemented as I move the
saving functionality into LGChangeHandler

Testing of just the smtpServer field was successful; this can be
expanded to the remaining text controls